# eagerly do boundscheck when indexing into CartesianIndices