--- a/base/multidimensional.jl
+++ b/base/multidimensional.jl
@@ -359,7 +359,7 @@ module IteratorsMD
     end
     @inline function Base.getindex(iter::CartesianIndices{N,R}, I::Vararg{Int, N}) where {N,R}
         # Eagerly do boundscheck before calculating each item of the CartesianIndex so that
-        # we can pass `@inbounds` hint to inside the map and generates more efficient SIMD codes
+        # we can pass `@inbounds` hint to inside the map and generates more efficient SIMD codes (#42115)
         @boundscheck checkbounds(iter, I...)
         index = map(iter.indices, I) do r, i
             @inbounds getindex(r, i)
@@ -370,13 +370,17 @@ module IteratorsMD
     # CartesianIndices act as a multidimensional range, so cartesian indexing of CartesianIndices
     # with compatible dimensions may be seen as indexing into the component ranges.
     # This may use the special indexing behavior implemented for ranges to return another CartesianIndices
-    @propagate_inbounds function Base.getindex(iter::CartesianIndices{N,R},
+    @inline function Base.getindex(iter::CartesianIndices{N,R},
         I::Vararg{Union{OrdinalRange{<:Integer, <:Integer}, Colon}, N}) where {N,R}
-        CartesianIndices(getindex.(iter.indices, I))
+        @boundscheck checkbounds(iter, I...)
+        indices = map(iter.indices, I) do r, i
+            @inbounds getindex(r, i)
+        end
+        CartesianIndices(indices)
     end
     @propagate_inbounds function Base.getindex(iter::CartesianIndices{N},
         C::CartesianIndices{N}) where {N}
-        CartesianIndices(getindex.(iter.indices, C.indices))
+        getindex(iter, C.indices...)
     end
 
     # If dimensions permit, we may index into a CartesianIndices directly instead of constructing a SubArray wrapper