--- a/base/multidimensional.jl
+++ b/base/multidimensional.jl
@@ -382,6 +382,7 @@ module IteratorsMD
         C::CartesianIndices{N}) where {N}
         getindex(iter, C.indices...)
     end
+    @inline Base.getindex(iter::CartesianIndices{0}, ::CartesianIndices{0}) = iter
 
     # If dimensions permit, we may index into a CartesianIndices directly instead of constructing a SubArray wrapper
     @propagate_inbounds function Base.view(c::CartesianIndices{N}, r::Vararg{Union{OrdinalRange{<:Integer, <:Integer}, Colon},N}) where {N}