This PR eagerly does boundscheck when doing `getindex(::CartesianIndices, ...)`, this generates simpler inlined codes and furthermore enables SIMD

```julia
julia> versioninfo()
Julia Version 1.8.0-DEV.473
Commit fce2daff16* (2021-09-04 09:37 UTC)
Platform Info:
  OS: macOS (x86_64-apple-darwin20.6.0)
  CPU: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
  WORD_SIZE: 64
  LIBM: libopenlibm
  LLVM: libLLVM-12.0.1 (ORCJIT, skylake)
```

```julia
function sumcart_oneto(X)
    out = zero(eltype(X))
    R = CartesianIndices(X) # OneTo
    @inbounds @simd for i in eachindex(X)
        out += X[R[i]]
    end
    return out
end

function sumcart_iur(X)
    out = zero(eltype(X))
    R = CartesianIndices(map(i->Base.IdentityUnitRange(1:i), size(X))) # IdentityUnitRange
    @inbounds @simd for i in eachindex(X)
        out += X[R[i]]
    end
    return out
end

function sumcart_unitrange(X)
    out = zero(eltype(X))
    R = CartesianIndices(map(i->1:i, size(X))) # UnitRange
    @inbounds @simd for i in eachindex(X)
        out += X[R[i]]
    end
    return out
end

X = rand(64, 64);

@btime sumcart_oneto($X);
# PR: 241.691 ns (0 allocations: 0 bytes)
# master: 10.526 μs (0 allocations: 0 bytes)
@btime sumcart_iur($X);
# PR: 236.788 ns (0 allocations: 0 bytes)
# master: 10.950 μs (0 allocations: 0 bytes)
@btime sumcart_unitrange($X);
# PR: 234.622 ns (0 allocations: 0 bytes)
# master: 10.529 μs (0 allocations: 0 bytes)
```

The earliest version that I obverse this performance regression is in Julia 1.4.0

A related benchmark:

```julia
function sumcart_linear(X)
    out = zero(eltype(X))
    @inbounds @simd for i in 1:length(X)
        out += X[i]
    end
    return out
end

@btime sumcart_linear($X);
# master: 219.778 ns (0 allocations: 0 bytes)
```

fixes #42115