@johnnychen94 2021-09-06
The `getindex(::CartesianIndices, ::RangeTypes)` methods are newly introduced in https://github.com/JuliaLang/julia/pull/40344 so I've removed them in d58ce6a. Doing this we can backport this PR to 1.6 and 1.7 branch. I'll submit a separate PR after this.